requirements.txt
fastapi
uvicorn
httpx
supabase
python-dotenv
free-proxy
python-multipart
asyncpg
